requirements.txt
feedparser
youtube-transcript-api
pyyaml
schedule
